Move `Client` traits to `client` module
