+++ examples/multi_middleware.rs
@@ -30,7 +30,7 @@ use std::net::SocketAddr;
 use std::process::Command;
 use std::time::Instant;
 
-use jsonrpsee::core::{middleware, traits::Client};
+use jsonrpsee::core::{client::Client, middleware};
 use jsonrpsee::rpc_params;
 use jsonrpsee::ws_client::WsClientBuilder;
 use jsonrpsee::ws_server::{RpcModule, WsServerBuilder};
