+++ examples/ws_sub_with_params.rs
@@ -26,7 +26,7 @@
 
 use std::net::SocketAddr;
 
-use jsonrpsee::core::traits::SubscriptionClient;
+use jsonrpsee::core::client::SubscriptionClient;
 use jsonrpsee::rpc_params;
 use jsonrpsee::ws_client::WsClientBuilder;
 use jsonrpsee::ws_server::{RpcModule, WsServerBuilder};
