+++ examples/ws_subscription.rs
@@ -26,8 +26,7 @@
 
 use std::net::SocketAddr;
 
-use jsonrpsee::core::client::Subscription;
-use jsonrpsee::core::traits::SubscriptionClient;
+use jsonrpsee::core::client::{Subscription, SubscriptionClient};
 use jsonrpsee::core::Error;
 use jsonrpsee::rpc_params;
 use jsonrpsee::ws_client::WsClientBuilder;
