+++ proc-macros/src/render_client.rs
@@ -41,9 +41,9 @@ impl RpcDescription {
 		let (impl_generics, type_generics, _) = self.trait_def.generics.split_for_impl();
 
 		let super_trait = if self.subscriptions.is_empty() {
-			quote! { #jsonrpsee::core::traits::Client }
+			quote! { #jsonrpsee::core::client::Client }
 		} else {
-			quote! { #jsonrpsee::core::traits::SubscriptionClient }
+			quote! { #jsonrpsee::core::client::SubscriptionClient }
 		};
 
 		let method_impls =
