+++ ws-client/src/tests.rs
@@ -29,8 +29,8 @@ use crate::types::error::{ErrorCode, ErrorObject, ErrorResponse};
 use crate::types::ParamsSer;
 use crate::WsClientBuilder;
 use jsonrpsee_core::client::Subscription;
+use jsonrpsee_core::client::{Client, SubscriptionClient};
 use jsonrpsee_core::rpc_params;
-use jsonrpsee_core::traits::{Client, SubscriptionClient};
 use jsonrpsee_core::Error;
 use jsonrpsee_test_utils::helpers::*;
 use jsonrpsee_test_utils::mocks::{Id, WebSocketTestServer};
